src/main/java/com/BridgeLabz/generics/MaximumAnalyser.java
package com.BridgeLabz.generics;

public class MaximumAnalyser {

    public Integer FindMaximum(Integer first, Integer second, Integer third) {
        if (first.compareTo(second) > 0 && first.compareTo(third) > 0)
            return first;
        else if (second.compareTo(first) > 0 && second.compareTo(third) > 0)
            return second;
        else
            return third;
    }
    public Float FindMaximum(Float first, Float second, Float third) {
        if (first.compareTo(second) > 0 && first.compareTo(third) > 0)
            return first;
        else if (second.compareTo(first) > 0 && second.compareTo(third) > 0)
            return second;
        else
            return third;
    }
}


